move run env script
